Aggiunta del JsonIgnoreProperties per eliminare le ripetizioni in chiamata

+++ backend/springboot/src/main/java/com/AssignmentTWEB/springboot/Actors/ActorRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ActorRepository extends JpaRepository<Actor, Integer> {
+public interface ActorRepository extends JpaRepository<Actor, Long> {
 
     List<Actor> findByMovie(Movie movie); //find all movie's actors
 }
