+++ backend/springboot/src/main/java/com/AssignmentTWEB/springboot/Posters/PostersRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PostersRepository extends JpaRepository <Posters, Integer>{
+public interface PostersRepository extends JpaRepository <Posters, Long>{
 
     //find all countries where a film has been produced
     Posters findByMovie(Movie movie);
