+++ backend/springboot/src/main/java/com/AssignmentTWEB/springboot/Studios/StudioRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface StudioRepository extends JpaRepository<Studio, Integer> {
+public interface StudioRepository extends JpaRepository<Studio, Long> {
 
     List<Studio> findByMovie(Movie movie); //find all movie's studios
